修改文件结构 为增添三个按钮：喜欢、讨论、分享
