--- src/components/quoteCard.css
@@ -1,28 +1,0 @@
-.card {
-    position: relative;
-    width: 40vw;
-    height: 30vh;
-    margin: 0 auto;
-    padding: 20px;
-    top: 20vh;
-    color: #fff;
-    border: 4px solid #ffffff;
-    border-radius: 10px;
-    box-shadow: 0px 50px 0px 0px #fff;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.mainContent {
-    font-family: 'Times New Roman', Times, serif;
-    font-size: 36px;
-}
-
-.contentSource {
-    position: absolute;
-    right: 20px;
-    bottom: -40px;
-    color: rgb(151, 151, 151);
-    font-family: 'Microsoft Sans Serif', sans-serif;
-}
