+++ src/components/quoteCard/quoteCard.css
@@ -1,0 +1,71 @@
+.card {
+    position: relative;
+    width: 50vw;
+    min-width: 400px;
+    max-height: 200px;
+    height: 30vh;
+    margin: 0 auto;
+    padding: 20px;
+    top: 20vh;
+    color: rgb(255, 255, 255);
+    border: 4px solid #ffffff;
+    border-radius: 12px;
+    box-shadow: 0px 50px 0px 0px #fff;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.mainContent {
+    font-family: 'Times New Roman', Times, serif;
+    font-size: 36px;
+}
+
+.contentSource {
+    color: rgb(151, 151, 151);
+    font-family: 'Microsoft Sans Serif', sans-serif;
+    font-weight: 500;
+}
+
+.text {
+    position: relative;
+}
+
+.stausBar {
+    position: fixed;
+    top: 53vh;
+    width: 50vw;
+    display: flex;
+    justify-content: space-between;
+}
+
+.stausBarButton {
+    width: 200px;
+    display: flex;
+    justify-content: space-between;
+}
+
+.textFavoriteButton {
+    color: rgb(255, 133, 133);
+    font-size: 28px;
+}
+
+.textFavoritedButton {
+    color: red;
+    font-size: 28px;
+}
+
+.textMessageButton {
+    color: rgb(170, 170, 170);
+    font-size: 28px;
+
+}
+
+.expand {
+    color: #000;
+}
+
+.textShareButton {
+    color: rgb(170, 170, 170);
+    font-size: 28px;
+}
